perf: 全面优化代码，去掉所有的开关判断
