+++ apps/libs/process/ssh.py
@@ -10,7 +10,7 @@ def should_kill(process):
     return process.pid != 1 and process.name() == 'ssh'
 
 
-def kill_ssh_children(pid):
+def kill_process_ssh_children(pid):
     """Kill all SSH child processes of a given PID."""
     try:
         process = psutil.Process(pid)
