+++ apps/ops/ansible/callback.py
@@ -2,8 +2,6 @@ import os
 from collections import defaultdict
 from functools import reduce
 
-from django.conf import settings
-
 
 class DefaultCallback:
     STATUS_MAPPER = {
@@ -162,12 +160,7 @@ class DefaultCallback:
     def status_handler(self, data, **kwargs):
         status = data.get('status', '')
         self.status = self.STATUS_MAPPER.get(status, 'unknown')
-
-        if settings.ANSIBLE_RECEPTOR_ENABLE:
-            self.private_data_dir = data.get("private_data_dir", None)
-        else:
-            rc = kwargs.get('runner_config', None)
-            self.private_data_dir = rc.private_data_dir if rc else '/tmp/'
+        self.private_data_dir = data.get("private_data_dir", None)
 
     def write_pid(self, pid):
         pid_filepath = os.path.join(self.private_data_dir, 'local.pid')
